[fingerprint_grow] Fix OOB write, uint16 overflow, and missing break
Fix out-of-bounds write in transfer_() when timeout occurs: the vector
is cleared at line 388 but then accessed via [0] at line 467. Use
push_back() instead.

Fix uint16_t overflow in aura_led_control(): 10 * speed * count can
reach 650,250 which overflows uint16_t. Change last_aura_led_duration_
to uint32_t.

Add missing break after ENROLL_MISMATCH case in save_fingerprint_() to
prevent unintended fall-through to default case.

+++ esphome/components/fingerprint_grow/fingerprint_grow.h
@@ -190,7 +190,7 @@ class FingerprintGrowComponent : public PollingComponent, public uart::UARTDevic
   bool is_sensor_awake_ = false;
   uint32_t last_transfer_ms_ = 0;
   uint32_t last_aura_led_control_ = 0;
-  uint16_t last_aura_led_duration_ = 0;
+  uint32_t last_aura_led_duration_ = 0;
   uint16_t system_identifier_code_ = 0;
   uint32_t idle_period_to_sleep_ms_ = UINT32_MAX;
   sensor::Sensor *fingerprint_count_sensor_{nullptr};
